forgot a cast

+++ hypothesis-python/src/hypothesis/core.py
@@ -1723,7 +1723,9 @@ def custom_mutator(data, buffer_size, seed):
                         min_val, max_val, smallest_nonzero_mag, random=random
                     )
             # with probability 0.05, truncate to an integer-valued float
-            if random.randint(0, 99) < 5 and permitted(truncated := math.floor(forced)):
+            if random.randint(0, 99) < 5 and permitted(
+                truncated := float(math.floor(forced))
+            ):
                 forced = truncated
         else:
             raise Exception(f"unhandled case ({ir_type=})")
